Move error to components
