+++ frontend/src/components/Error.vue
